+++ src/packetforge-ng.c
@@ -533,7 +533,7 @@ static int my_encrypt_data(unsigned char * dest,
 		}
 	}
 
-	if (opt.prgalen - 4 < length)
+	if (opt.prgalen - 4 < (size_t) length)
 	{
 		printf(
 			"Please specify a longer PRGA file (-y) with at least %i bytes.\n",
